Removed Double data type due to low performance

+++ include/core/Hashtable.h
@@ -159,19 +159,19 @@ namespace jm
          HashtableEntry** mData;
 
           //! \brief This variable stores the actual length of the storage array.
-         Integer mArrLength;
+         int64 mArrLength;
 
           //! \brief This variable stores the number of entries in the hashtable.
-         Integer mDataLength;
+         int64 mDataLength;
 
           //! \brief This threshold defines the number of entries at which the storage array for 
           //! the data should be resized. It is the absolute number of entries at which a storage
           //! reorganization is performed.
-         Integer mThreshold;
+         int64 mThreshold;
 
           //! This variable stores the load factor threshold at which the Hashtable should be rehashed.
           //! A value of 0 means 0%, while a value of 1 means 100%.
-         Double mLoadfactor;
+         double mLoadfactor;
 
           /*!
            \brief This subclass implements an iterator that allows iterating through the hashtable.
@@ -205,7 +205,7 @@ namespace jm
                Hashtable* table;
                HashtableEntry* entry;
                HashtableEntry* last;
-               Integer index;
+               int64 index;
          };
 
    };
